+++ tsdoc/src/index.ts
@@ -1,6 +1,4 @@
 
-export { DocNodeKind, DocNode } from './api/nodes/DocNode';
-export { DocComment } from './api/nodes/DocComment';
-
-export { TSDocParser } from './api/TSDocParser';
-export { TextRange, ITextLocation } from './api/TextRange';
+export * from './nodes';
+export { TSDocParser } from './parser/TSDocParser';
+export { TextRange, ITextLocation } from './parser/TextRange';
