+++ tsdoc/src/nodes/DocError.ts
