+++ tsdoc/src/nodes/DocPlainText.ts
