+++ tsdoc/src/parser/LineExtractor.ts
