+++ tsdoc/src/parser/ParseError.ts
