+++ tsdoc/src/parser/TextRange.ts
